feat: 토너먼트 순위 UI 구현
